+++ appGUI/preferences/gerber/GerberPreferencesUI.py
@@ -25,7 +25,7 @@ class GerberPreferencesUI(QtWidgets.QWidget):
         self.defaults = defaults
 
         self.gerber_gen_group = GerberGenPrefGroupUI(decimals=self.decimals, defaults=self.defaults)
-        self.gerber_gen_group.setMinimumWidth(250)
+        self.gerber_gen_group.setMinimumWidth(200)
         self.gerber_opt_group = GerberOptPrefGroupUI(decimals=self.decimals, defaults=self.defaults)
         self.gerber_opt_group.setMinimumWidth(250)
         self.gerber_exp_group = GerberExpPrefGroupUI(decimals=self.decimals, defaults=self.defaults)